tests/wpt/tests/css/css-font-loading/historical.html

<!doctype html>
<meta charset=utf-8>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<link rel=help href="https://drafts.csswg.org/css-font-loading/">

<div id=log></div>
<script>
"use strict";

// https://github.com/w3c/csswg-drafts/commit/b03a8c9721b56774723ed7804ca6f3f04a922d69
test(function() {
  assert_throws_js(TypeError, function() {
    new FontFaceSet([]);
  });
}, "FontFaceSet constructor should not be supported");
</script>
